+++ src/test/javascript/identifyjs/identifyAddressTest.js
@@ -163,6 +163,59 @@ describe("Identify address Test Suite: ValidateMailingAddressPremium", function(
 	}); 
 });
 
+
+/********************ValidateMailingAddressUSCan Test Cases ***************************************************/
+
+
+describe("Identify address Test Suite: ValidateMailingAddressUSCan", function(){
+	var IA;	
+	beforeEach(function() {
+		IA = new IDENTIFYAPIS.identifyAddress('<access token>'); //PUT TOKEN HERE
+	});
+    it("Identify Address test json validate address USCan", function () {
+        var identifyAddress = IA.validateMailingAddressUSCan('{ "options": { "OutputCasing":"M" }, "Input": { "Row": [ { "AddressLine1": "1 Global View", "AddressLine2": "", "City": "Troy", "Country": "us", "StateProvince": "NY", "PostalCode": "","FirmName":"Pitney Bowes Software", "user_fields":[{"name":"name1","value":"value1"},{"name":"name2","value":"value2"}] } ] } }');
+        expect(identifyAddress).not.toBeNull();
+		expect(identifyAddress.httpResponse.status).toEqual(200);
+    }); 
+	it("Identify Address test json empty validate address USCan", function () {
+        expect( function(){ IA.validateMailingAddressUSCan(''); } ).toThrow(new Error("PostData is of invalid type, it should be either json or xml."));
+    });
+	it("Identify Address test json batch validate address USCan", function () {
+		var identifyAddress = IA.validateMailingAddressUSCan('{ "options": { "OutputCasing":"M" }, "Input": { "Row": [ { "AddressLine1": "50 water st", "AddressLine2": "", "City": "lee", "Country": "us", "StateProvince": "ma", "PostalCode": "","user_fields":[{"name":"name1","value":"value1"},{"name":"name2","value":"value2"}] },{ "AddressLine1": "1 Global View", "AddressLine2": "", "City": "Troy", "Country": "us", "StateProvince": "ny", "PostalCode": "","FirmName":"Pitney Bowes Software"} ] } }');
+		expect(identifyAddress).not.toBeNull();
+		expect(identifyAddress.httpResponse.status).toEqual(200);
+	});
+	it("Identify Address test xml validate address USCan", function () {
+        var identifyAddress = IA.validateMailingAddressUSCan('<ValidateMailingAddressUSCANAPIRequest> <options> <OutputCasing>U</OutputCasing> </options> <Input> <Row> <PostalCode>01238</PostalCode> <StateProvince>NY</StateProvince> <AddressLine1>1 Global View</AddressLine1> <City>Troy</City> <AddressLine2/> <Country>US</Country> <FirmName>Pitney Bowes Software</FirmName> </Row> </Input> </ValidateMailingAddressUSCANAPIRequest>');
+        expect(identifyAddress).not.toBeNull();
+		expect(identifyAddress.httpResponse.status).toEqual(200);
+    });
+	it("Identify Address test xml batch validate address USCan", function () {
+        var identifyAddress = IA.validateMailingAddressUSCan('<ValidateMailingAddressUSCANAPIRequest> <options> <OutputCasing>U</OutputCasing> </options> <Input> <Row> <PostalCode>01238</PostalCode> <StateProvince>NY</StateProvince> <AddressLine1>1 Global View</AddressLine1> <City>Troy</City> <AddressLine2/> <Country>US</Country> <FirmName>Pitney Bowes Software</FirmName> </Row> <Row> <PostalCode></PostalCode> <StateProvince>ma</StateProvince> <AddressLine1>50 water st.</AddressLine1> <City>lee</City> <AddressLine2/> <Country>US</Country> </Row> </Input> </ValidateMailingAddressUSCANAPIRequest>');
+        expect(identifyAddress).not.toBeNull();
+		expect(identifyAddress.httpResponse.status).toEqual(200);
+    });
+ 
+	it("Identify Address valid batch parameters - validate address USCan", function () {
+		expect( function(){ IA.validateMailingAddressUSCan('Neither xml nor json'); } ).toThrow(new Error("PostData is of invalid type, it should be either json or xml."));
+    });
+});
+
+describe("Identify address Test Suite: validateMailingAddressUSCan", function(){
+	var IA;	
+	beforeEach(function() {
+		IA = new IDENTIFYAPIS.identifyAddress('JUNK');//testing invalid token
+	});
+    it("validateMailingAddressUSCan test invalid token json", function () {
+        var identifyAddress = IA.validateMailingAddressUSCan('{ "options": { "OutputCasing":"M" }, "Input": { "Row": [ { "AddressLine1": "1 Global View", "AddressLine2": "", "City": "Troy", "Country": "us", "StateProvince": "NY", "PostalCode": "","FirmName":"Pitney Bowes Software", "user_fields":[{"name":"name1","value":"value1"},{"name":"name2","value":"value2"}] } ] } }');
+		expect(identifyAddress.httpResponse.status).toEqual(401);
+    }); 
+	it("validateMailingAddressUSCan test invalid token xml", function () {
+	var identifyAddress = IA.validateMailingAddressUSCan('<ValidateMailingAddressUSCANAPIRequest> <options> <OutputCasing>U</OutputCasing> </options> <Input> <Row> <PostalCode>01238</PostalCode> <StateProvince>NY</StateProvince> <AddressLine1>1 Global View</AddressLine1> <City>Troy</City> <AddressLine2/> <Country>US</Country> <FirmName>Pitney Bowes Software</FirmName> </Row> </Input> </ValidateMailingAddressUSCANAPIRequest>');
+	expect(identifyAddress.httpResponse.status).toEqual(401);
+	}); 
+});
+
 /********************GetCityStateProvince Test Cases ***************************************************/