+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ### Added
 - Added 4.0 Version of JavaScript SDK which has Identify API features IdentifyAddress, IdentifyRisk, IdentifyEmail and IdentifyExtract.
 - Added 4.1 Version of JavaScript SDK. Added GetCityStateProvinceAPI and GetPostalCodesAPI under IdentifyAddress.
+- Added 4.2 Version of JavaScript SDK. Added ValidateMailingAddressUSCANAPI under IdentifyAddress.